server/requirements.txt
fastapi==0.136.1
motor==3.7.1
pydantic==2.13.3
pymongo==4.17.0
python-dotenv==1.2.2
python_jose==3.5.0
